DRY'ing up name change handlers
