+++ web/src/client/js/components/FieldNumber/FieldNumberComponent.js
@@ -4,19 +4,13 @@ import PropTypes from 'prop-types'
 import { debounce } from 'Shared/utilities'
 import './FieldNumber.scss'
 
-const FieldNumberComponent = ({ field, onChange, showName }) => {
-  // field name change
-  const nameRef = useRef()
-  const debouncedNameChangeHandler = debounce(1000, (e) => {
-    onChange(field.id, { name: nameRef.current.value })
-  })
+const FieldNumberComponent = ({ field, onChange }) => {
   // Value change
   const valueRef = useRef()
   const debouncedValueChangeHandler = debounce(1000, () => {
     onChange(field.id, { value: valueRef.current.value })
   })
   return (
-    <>
     <input
       className="h-second-level"
       defaultValue={field.value}
@@ -24,23 +18,12 @@ const FieldNumberComponent = ({ field, onChange, showName }) => {
       onFocus={(e) => { e.target.select() }}
       type="number"
       ref={valueRef} />
-    {showName &&
-    <input
-      ref={nameRef}
-      className="field__name"
-      defaultValue={field.name}
-      onChange={debouncedNameChangeHandler}
-      onFocus={(e) => { e.target.select() }}
-      type="text" />
-    }
-    </>
   )
 }
 
 FieldNumberComponent.propTypes = {
   field: PropTypes.object.isRequired,
-  onChange: PropTypes.func.isRequired,
-  showName: PropTypes.bool
+  onChange: PropTypes.func.isRequired
 }
 
 export default FieldNumberComponent
